extract and split up image and video examples for composable container
